projectile/projectile_engine-main/requirements.txt
streamlit
numpy
pandas
plotly
scipy
torch
joblib
scikit-learn
google-generativeai
openai
watchdog
